Move the unit tests from commandlinerunner to separate project
See gh-142
