+++ boot/tcf/src/test/java/com/example/boot/tcf/ArgsTest.java
@@ -1,4 +1,4 @@
-package com.example.commandlinerunner;
+package com.example.boot.tcf;
 
 import java.util.List;
 
@@ -7,19 +7,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.UseMainMethod;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(webEnvironment = WebEnvironment.NONE, useMainMethod = UseMainMethod.ALWAYS, args = "--test.args=1")
-class CommandlinerunnerApplicationTests {
-
-	@Test
-	void mainShouldRun() {
-		String property = System.getProperty("main.ran", "false");
-		assertThat(property).isEqualTo("true");
-	}
+@SpringBootTest(webEnvironment = WebEnvironment.NONE, args = "--test.args=1")
+class ArgsTest {
 
 	@Test
 	void shouldPassArgs(@Autowired ApplicationArguments applicationArguments) {
