+++ boot/tcf/src/test/java/com/example/boot/tcf/OutputCaptureTests.java
@@ -1,4 +1,4 @@
-package com.example.commandlinerunner;
+package com.example.boot.tcf;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
